+++ Faker_currency_symbol.v
@@ -1,0 +1,5 @@
+module faker
+
+pub fn (mut this Faker) currency_symbol() string {
+    return this.random_element(Currency.cases()).to_symbol()
+}