+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Company name fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 - Visa credit card number fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 - MAC address fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
+- Currency symbol fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
+- Currency code fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 
 ### Fixed